feat(mobile): change how we deal with .env variables
Previously we served a .env.development file which would be automatically used by expo when generating the app. I’ve renamed it
to .env.example and users are supposed to create an .env.local file themselves.
Also added a new variable GOOGLE_SERVICES_JSON_DEV that needs to be provided when building the dev variant of the app.