fix(mobile): unable to subscribe to push notifications on android
When calling the register/notifications endpoint on android we were recieving a 422. It turned out that this is due to the format of the uuid that we send. The react-native-device-crypto getUniqueId method return uuid v4 on iOS, but only a 64-bix hex on android. Now, I’m converting the android id to a unique uuid. This way we are now able to subscribe on the gateway.