fix(mobile): add better detection for strongbox support
Despite strongbox support in SDK v28 there are a lot of devices that don’t have the hardware component to support strongbox encryption. Relying solely on the sdk version is not good enough. We try to store a key with strongbox if that fail we know that this is not supported and don’t turn it on.