fix(mobile): signer select list empty on android
The way we’ve overriden the headers of fetch wa causing problems on Android. We were creating a new object instead of a Header Object and it seems that Android is very picky when it comes to the structure of the fetch header.
It is interesting why it only happens for eth_getBalance calls, and for nothing else, but switchting to a Headers object resolves the issue.