feat(store): add hooks into the cgwClient lifecycle
The Auth endpoint returns a response cookie. This cookie is automatically set by browsers, but this doesn’t work that way in React-Native context. Because of this we need to be able to intercept the requests and deal with the cookie ourselves. By providing those hooks (or „escape hatches“), we can hook into the request process and manage the hooks on the mobile side.