fix(ExecTransaction): replace Summary component with TxData
For an execTransaction call we don't have all the data required to display the full nested transaction with the advanced tx details. That is why e.g. the nonce is showing as -1 and when switching to the JSON tab in the receipt, we get an error.
For that reason, we update the nestedTransaction display for execTransaction calls to use the TxData component instead of Summary.