refactor(mobile): change safe slice
The safeSlice was storing an aggragated fiatbalance, but it was containing the safeInfo of one of the deployments of the safe. In a multisafe setup it meant that the safeInfo stored would be for example mainnet, yet we could be showing the sepolia deployment on the screen. This lead to subtle errors such as the threshold being wrong in some places of the app.

The rafactoring now makes sure that we store all versions of safeInfo inthe safesSlice. We also have an extraReducer that is triggered whenever the safesGetOverviewForMany query runs. This way we can keep the state of the safeInfo always up to date.