fix(mobile): use generic biometrics term
On android there are devices that have both fingerprint and face id. Our detection does not detect which one of them is enabled but rather what is available. There could be a case where a user has enabled touch id but we show face id in the ui. That is why we go for the generic biometrics term.