fix(mobile): properly reset navigation state
The navigation on the add signer screen grew quite complex. We could enter from multiple screens, but at the end the expected result is tha the user is sent to the “assets” screen. By reseting the navigation state we can be sure that the user will end up on the tabs. Dimissing screens, going back was fragile as we never knew exactly how many screens where in the stack.