fix(mobile): advanced details crashes when no txData
Some transactions such as “sent” events due to a swap don’t have txData. For those on web we don’t display any advanced details. Now we do the same. I also added a check to formatTxDetails for txData so that it won’t crash again if we stumble on a tx without any txData.