fix(mobile): sorting for my accounts not working
If we had more accounts visible on the screen, than the available screen height we weren’t able to scroll. I’m not 100% sure how the individual views are stacking and who is strealing the drag event, but if we switch to just having a DraggableFlatList without wrapping it in a view it works fine.