fix(mobile): disable default android backup
Android is trying to backup the local database. Per se this is not an issue, but we use a flag in this database that indicates whether a PK was imported. If you restore such backup it would look like you have a PK imported, but you won’t be able to use this key as it is not there. That’s why I’ve opted to disable backup all together.