fix(mobile): importing safe was not showing network errors
The import safe function was caching the network request and was not redoing it. So if the user had bad internet connection at the time of import he would get a network request error - this request would be cached and we would always return empty data. We would not display any error - so the user wouldn’t know what was going on. Now on error we display a “network error” message. In addition to that we dispatch the network request again if the user modifies the input.