fix(mobile): persist parts of RTK’s slice
Our cgwClientFilter is actually persisting parts of RTK’s slice. Unfortunately in https://github.com/safe-global/safe-wallet-monorepo/pull/6348 we erronously added cgwClient.reducerPath to the blacklist filter.
This can have unpleasant effects if the user starts the app and he has a network error - it might seem that we have data, but the chains are missing.