chore: unify prettier setup across repo
there were slight inconsistencies in the prettier setup across the repo which lead to annoying problems. Running prettier from the root would modify files in subdirectories. But if you would run prettier from those subdirecotry -> the same files that were previously modified wouldn’t be treated as a problem.

Now every prettier command across the repo is loading the same root prettier config. In addition to this lintstaged should now only run on the staged files and not on the complete code base.